Attempt to tackle with mineral concentration calculation
06.23.2023

In RandomMineralMap,
- new: add createMinerals(), calculateIteration()
- change: rework getAllMineralConcentrations(),
          getMineralConcentration()
- change: rename and rework createLocalConcentration()
- change: comment out getMineralConcentrationEffect()

In MineralMapLayer,
- change: by default, each mineral will need to be selected
          to be viewable in and out surface map
- change: check if minerals are selected in
          determineMineralConcentrations()

Relates to #935 and #942

+++ mars-sim-console/src/main/java/org/mars/sim/console/chat/simcommand/ExploredCommand.java
@@ -54,8 +54,8 @@ public class ExploredCommand extends ChatCommand {
 
 				// Check the locaiton has minerals
 				boolean hasMinerals = false;
-				Map<String,Double> minerals = surface.getMineralMap().getAllMineralConcentrations(siteLocation);
-				for(Double conc : minerals.values()) {
+				Map<String, Integer> minerals = surface.getMineralMap().getAllMineralConcentrations(siteLocation);
+				for(int conc : minerals.values()) {
 					hasMinerals = conc > 0D;
 					if (hasMinerals) {
 						break;
