+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/mission/MiningMissionCustomInfoPanel.java
@@ -175,7 +175,7 @@ public class MiningMissionCustomInfoPanel extends MissionCustomInfoPanel {
 
 		// Data members.
 		protected Map<String, Double> estimatedConcentrationMap;
-		protected Map<String, Double> actualConcentrationMap;
+		protected Map<String, Integer> actualConcentrationMap;
 
 		/**
 		 * Constructor
